Refactoring of insecure connection management
- still allow secure connections on localhost without credentials
- `create_insecure_channel_and_connect` function now imported directly instead of being patched
- minor documentation addition
